[ZEPPELIN-6457] Extract shared Python ZeppelinContext helpers
