+++ spark/interpreter/src/main/resources/python/zeppelin_ipyspark.py
@@ -64,19 +64,7 @@ if intp.isAfterSpark33():
 else:
     sqlContext = sqlc = __zSqlc__ = __zSpark__._wrapped
 
-class IPySparkZeppelinContext(PyZeppelinContext):
-
-    def __init__(self, z, gateway):
-        super(IPySparkZeppelinContext, self).__init__(z, gateway)
-
-    def show(self, obj, **kwargs):
-        from pyspark.sql import DataFrame
-        if isinstance(obj, DataFrame):
-            print(self.z.showData(obj._jdf))
-        else:
-            super(IPySparkZeppelinContext, self).show(obj, **kwargs)
-
-z = __zeppelin__ = IPySparkZeppelinContext(intp.getZeppelinContext(), gateway)
+z = __zeppelin__ = PySparkZeppelinContext(intp.getZeppelinContext(), gateway)
 
 # add jars to path
 import sys
